+++ types.go
@@ -1,5 +1,11 @@
 package main
 
+// ListNode is a list node
+type ListNode struct {
+	Val  int
+	Next *ListNode
+}
+
 // TreeNode is a binary tree node.
 type TreeNode struct {
 	Val   int