+++ CHANGELOG.md
@@ -513,6 +513,10 @@ $$Pow({x}, {n})=Pow\left({x}^2, \dfrac{n}{2}\right) * Pow\left(x, n \mod 2\right
 
 - Recursion
 
+## 0109 - Convert Sorted List to Binary Search Tree
+
+- Recursion
+
 ## 0415 - Add strings
 
 - Full adder